Lay groundwork for argument rebind support
For argument rebind to be possible, every SQL element with the method
AppendSQL/AppendSQLExclude must take in an extra argument called
'param'. param is a map[string]int that maps a param name to the index
of an argument in the args slice. When the user wants to provide a
argument that can be rebinded later they simply have to pass in a
placeholder param i.e. sq.Param(name, defaultValue). To rebind that
argument we do a name lookup in the param map to find the index of the
argument in the args slice, then swap it out for the new argument.