Added panic recovery to Insert/Update
Validation logic can be added to ColumnMappers, and validation errors
are propagated by panicking with the errors (which are recovered in the
Exec() method).  ColumnMappers appear in InsertQuery and UpdateQuery so
they must now incorporate panic recovery in order to recover error
panics and return them as errors.

The downside is that ToSQL and AppendSQL do not have any error returns
in their function signatures. I don't intend on adding an error return
to their function signatures as I think it was the right choice: the
query builder has no reason to give an error -unless- the user writes
custom validation code in the ColumnMapper. If the user does write
custom validation code in the ColumnMapper, my workaround is that ToSQL
will recover from any panics and stuffs the panicked value into the args
slice. It also sets the query string to an empty string to indicate that
a panic occurred. The main loser of this design is AppendSQL which is no
longer safe to call as it may panic if the ColumnMapper panics. My
solution is to just document to users that AppendSQL for
InsertQuery/UpdateQuery are not safe to call as they may panic, use
ToSQL instead.

Ultimately I'm unsatisfied with how using validation logic normalises
writing panics. This is a decision I may have to revisit in the future.