Unexported irrelevant functions
the Row* family of functions already exist as methods on the Row struct,
I'm not sure why I added them in the first place. In addition they would
be outdated by the time Go Generics arrive since a single generic function
would be able to take their place. So I'm unexporting them in
preparation for that.