Prevent byte slices getting slice-expanded
Slice expansion should apply to every other slice type *except* for byte
slices, because they are considered a value on their own and not simply
a collection of disparate bytes.

Also included is a tweak of row.ScanInto to use reflection directly to
set the dest pointer. The prior version was inefficiently abusing
database/sql's (*Row).Scan() method over and over to set the dest
pointer, simply because I did not know how to use the reflect library at
that time.