Missing functionality on ArrayField (#10)
* Do not exapand slices implementing driver.Value

This is used when an array of custom enum in Go code is
stored in a database column of text[] type. Only applicable
for postgresql.

* Added ability to set array field to another field

This functionality is needed when an ArrayField must
be mapped to other field instead of a value. With regular
fields, the `Set()` function accepts any value so this
special case is only needed for ArrayField.

* Changed ArrayField.Set() to accept any value

This aligns ArrayField.Set() function with other field
definitions. This should be a backwards compatible
change as we are relaxing the type requirement.

* Fixed ArrayField.Set() to properly process slices

If user passes a slice to ArrayField.Set() it must be
converted to ArrayField. Otherwise, the slice will not
be properly represented as an array in the expanded sql.